vita-mcp/requirements.txt
aorta-sirius-dev
fastmcp
mcpo